#947 - `mars-sim-tools` created, and Mapping classes consolidated to `mars-sim-mapdata`
