+++ mars-sim-mapdata/pom.xml
@@ -12,6 +12,11 @@
 	<name>mars-sim-mapdata</name>
 	<dependencies>
 		<dependency>
+			<groupId>com.github.mars-sim</groupId>
+			<artifactId>mars-sim-tools</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+		<dependency>
 		    <groupId>de.articdive</groupId>
 		    <artifactId>jnoise-pipeline</artifactId>
 		    <version>4.0.0</version>
