+++ mars-sim-mapdata/src/main/java/org/mars_sim/msp/ui/swing/tool/map/MapLayer.java
@@ -1,0 +1,27 @@
+/*
+ * Mars Simulation Project
+ * MapLayer.java
+ * @date 2022-08-02
+ * @author Scott Davis
+ */
+
+package org.mars_sim.msp.ui.swing.tool.map;
+
+import java.awt.Graphics;
+
+import org.mars_sim.msp.core.Coordinates;
+
+
+/**
+ * The MapLayer interface is a graphics layer painted on the map display.
+ */
+public interface MapLayer {
+	/**
+	 * Displays the layer on the map image.
+	 * 
+	 * @param mapCenter the location of the center of the map.
+	 * @param baseMap   the base map controlling coordinate frame
+	 * @param g         graphics context of the map display.
+	 */
+	public void displayLayer(Coordinates mapCenter, Map baseMap, Graphics g);
+}
