+++ mars-sim-tools/pom.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>com.github.mars-sim</groupId>
+    <artifactId>mars-sim</artifactId>
+    <version>pre-3.6.0</version>
+  </parent>
+  <artifactId>mars-sim-tools</artifactId>
+</project>
